Added location update distance parameter

+++ include/cfg_entries.h
@@ -8,7 +8,7 @@ struct CfgEntries
     String deviceId;
     String authToken;
     uint32_t workDelay;
-    uint32_t updateDist;
+    uint16_t updateDist;
 };
 
 #endif // _CFG_ENTRIES_H_
