+++ include/config.h
@@ -15,7 +15,7 @@ const uint32_t USB_SERIAL_SPEED = 9600;
 
 const uint32_t SIM_SERIAL_SPEED = 9600;
 
-const uint32_t LOCATION_UPD_DIST = 0;
+const uint16_t LOCATION_UPD_DIST = 0;
 
 HardwareSerial & SIM_SERIAL = Serial3;
